refactor: move download history tests to dedicated package
- Moved download_history_test.go from synology_drive_exporter to
  download_history package
- All test functionality remains intact, just in a more appropriate location
